requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
slack-bolt==1.20.0
slack-sdk==3.32.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.1.4
langchain-community==0.3.0
openai==1.50.0
chromadb==0.5.15
apscheduler==3.10.4
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
bcrypt==4.2.0
python-multipart==0.0.12
httpx==0.27.2
pypdf==5.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
aiofiles==24.1.0